// client/src/components/styled-components/component-interfaces.ts
export interface IBox {
    border?: string
    width?: string
    borderRadius?: string
}

export interface ISearchDiv {
    display?: string
}

export interface ITitle {
    size?: string
    marginTop?: string
}

export interface IImage {
    url: string
}

export interface IUserBox {
    padding?: string
}

export interface IIcon {
    url: string
    urlHover: string
}

export interface IInput {
    error: boolean
}