// client/src/components/styled-components/Button.ts
import styled from 'styled-components'

export const Button = styled.button`
    background-color: #3B49DF;
    border: none;
    border-radius: 5px;
    color: #FFF;
    cursor: pointer;
    font-weight: 600;
    transition-duration: .3s;
    font-family: 'Inconsolata', monospace, Arial, Helvetica, sans-serif;

    &:hover {
        background-color: #323EBE;
    }
`